实现一个简单的注册接口

+++ train-member/src/main/java/com/example/train/member/controller/MemberController.java
@@ -23,7 +23,7 @@ public class MemberController {
     }
 
     @PostMapping("/register")
-    public int register(String mobile) {
+    public long register(String mobile) {
         return memberService.register(mobile);
     }
 }
